refactor(app): rename mapview folder
