+++ src/ClientApp/src/App.js
@@ -4,11 +4,11 @@ import { ErrorBoundary } from 'react-error-boundary';
 import { useImmerReducer } from 'use-immer';
 import './App.css';
 import { FallbackComponent } from './components/ErrorBoundary';
-import MapView from './components/esrijs/MapView';
 import { Filters } from './components/Filters';
 import Header from './components/Header';
 import { IdentifyContainer, IdentifyInformation } from './components/Identify';
 import MapLens from './components/MapLens';
+import MapView from './components/Mapping/MapView';
 import Sidebar from './components/Sidebar';
 import { mappingConfig } from './config';
 import UserContext from './UserContext';
